Distinguish better the empty error from real ones
"An empty string is an error we should not display" is too
unclean/fragile (it's much easier to miss, for instance): it's an Option
which is not statically tracked (and therefore, not documented). Use a
real Option instead.

Thanks to the previous commit, which encapsulated to some extent the
representation of errors, this change is relatively small.